perf: evaluate source lazily
