+++ packages/rspack/src/util/index.ts
@@ -75,17 +75,3 @@ export function asArray<T>(item: T): T[];
 export function asArray<T>(item: T | T[]): T[] {
 	return Array.isArray(item) ? item : [item];
 }
-
-export function toJsAssetInfo(info?: AssetInfo): JsAssetInfo {
-	return {
-		immutable: false,
-		minimized: false,
-		development: false,
-		hotModuleReplacement: false,
-		related: {},
-		chunkHash: [],
-		contentHash: [],
-		version: "",
-		...info
-	};
-}
