+++ packages/rspack/src/values/asset.ts
@@ -1,0 +1,16 @@
+import type { JsAssetInfo } from "@rspack/binding";
+import { AssetInfo } from "../Compilation";
+
+export function toJsAssetInfo(info?: AssetInfo): JsAssetInfo {
+	return {
+		immutable: false,
+		minimized: false,
+		development: false,
+		hotModuleReplacement: false,
+		related: {},
+		chunkHash: [],
+		contentHash: [],
+		version: "",
+		...info
+	};
+}
