+++ packages/rspack/src/values/index.ts
@@ -1,0 +1,2 @@
+export * as source from "./source";
+export * as asset from "./asset";
